requirements.txt
awscli==1.11.2
awsebcli==3.7.8
blessed==1.9.5
boto==2.42.0
botocore==1.4.60
cement==2.8.2
colorama==0.3.7
Django==1.10.2
django-appconf==1.0.2
django-compressor==2.1
django-sass-processor==0.5.1
django-storages==1.5.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
google-api-python-client==1.5.4
httplib2==0.9.2
jmespath==0.9.0
libsass==0.11.1
oauth2client==4.0.0
pathspec==0.3.4
Pillow==3.4.2
psycopg2==2.6.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.5.3
PyYAML==3.12
rcssmin==1.0.6
requests==2.9.1
rjsmin==1.0.12
rsa==3.4.2
s3transfer==0.1.7
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
uritemplate==3.0.0
wcwidth==0.1.7
websocket-client==0.37.0
